adding page to table cancel clean pass input in insert scan adding Js file to prevent row from closing after grid update adding preprepo with no actions or success adding OnFetch in jsworker to updaye client asap! adding K_Order class adding K_Order Controller

+++ HashtilERP/Shared/Models/K_Passport.cs
@@ -9,7 +9,7 @@ namespace HashtilERP.Shared.Models
     public class K_Passport
     {
         [Key]
-        public long K_PassportId { get; set; }
+        public int K_PassportId { get; set; }
 
         public int? SapDocEntry { get; set; }
         public int? PassportNum { get; set; }
